+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Added
   * Add Liveness and readiness probes
   * Add Dns Name Label Reuse Policy
+  * Add Identity Block
 
 # v7.3.0 - 2022-12-09